android app

+++ src/main/java/ru/litvinov/patientnotificator/service/ExolveSchedulerService.java
@@ -1,12 +1,10 @@
 package ru.litvinov.patientnotificator.service;
 
 
-import jakarta.annotation.PostConstruct;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 import ru.litvinov.patientnotificator.model.*;
 import ru.litvinov.patientnotificator.repository.SchedulerTaskRepository;
@@ -24,9 +22,9 @@ import java.util.stream.Collectors;
 import static ru.litvinov.patientnotificator.util.Constants.NAME_TAG;
 
 @Slf4j
-@Service
+//@Service
 @RequiredArgsConstructor
-public class SchedulerService {
+public class ExolveSchedulerService implements ISchedulerService {
 
     @Value("${exolve.sms.phone}")
     private String phone;
@@ -45,7 +43,8 @@ public class SchedulerService {
 
     private final SmsService smsService;
 
-    @PostConstruct
+    @Override
+//    @PostConstruct
     public void init() {
         schedulerTaskRepository.findAll().forEach(st -> {
             st.setFuture(submit(st));
@@ -55,6 +54,7 @@ public class SchedulerService {
         log.info("tasks in cache: {}", taskCache.stream().map(SchedulerTask::toString).collect(Collectors.joining("\n")));
     }
 
+    @Override
     public void schedule(final Patient patient) {
         final List<SchedulerTask> allByPatientAndLayoutType = schedulerTaskRepository.findAllByPatientAndLayout_Type(patient, Layout.Type.QUESTION);
         if (!CollectionUtils.isEmpty(allByPatientAndLayoutType)) {
@@ -68,6 +68,7 @@ public class SchedulerService {
         taskCache.add(schedulerTask);
     }
 
+    @Override
     public void schedule(final Patient patient, final LocalDateTime executionDate, final Layout layout) {
         final var schedulerTask = new SchedulerTask(UUID.randomUUID(), executionDate, layout, patient);
         schedulerTaskRepository.saveAndFlush(schedulerTask);
@@ -75,7 +76,8 @@ public class SchedulerService {
         taskCache.add(schedulerTask);
     }
 
-    private ScheduledFuture<?> submit(final SchedulerTask schedulerTask) {
+    @Override
+    public ScheduledFuture<?> submit(final SchedulerTask schedulerTask) {
         final Runnable runnable = () -> {
             log.info("executing task {}", schedulerTask);
             final var patient = schedulerTask.getPatient();
@@ -88,6 +90,7 @@ public class SchedulerService {
         return executorService.schedule(runnable, ChronoUnit.MINUTES.between(LocalDateTime.now(), schedulerTask.getExecutionDate()), TimeUnit.MINUTES);
     }
 
+    @Override
     public void deleteAllByPatient(final Patient patient) {
         final var allByPatient = schedulerTaskRepository.findAllByPatient(patient);
         if (CollectionUtils.isEmpty(allByPatient)) return;
@@ -98,7 +101,9 @@ public class SchedulerService {
         schedulerTaskRepository.deleteAll(allByPatient);
     }
 
+    @Override
     public void delete(final SchedulerTask schedulerTask) {
+        if (Objects.isNull(schedulerTask)) return;
         Optional.ofNullable(schedulerTask.getFuture()).ifPresent(sf -> sf.cancel(true));
         taskCache.remove(schedulerTask);
         schedulerTaskRepository.delete(schedulerTask);
